fix bookmark form file structure
